django-backend/requirements.txt
# Django и основные зависимости
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# База данных
psycopg2-binary==2.9.9

# Переменные окружения
python-dotenv==1.0.0

# HTTP клиенты
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2

# Playwright для аутентификации
playwright==1.40.0

# Парсинг и обработка данных
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2

# Утилиты
python-dateutil==2.8.2
pytz==2023.3

# Валидация
email-validator==2.1.0

